requirements.txt
flask==3.1.0
numpy==2.2.3
faiss-cpu==1.10.0
sentence-transformers==3.4.1
pygments==2.19.1
chardet==5.2.0
torch==2.6.0

